requirements.txt
cassandra-driver
lz4
scales